splitted everything in files, switch not working yet
